Remove dummy data entiresly from models [finishes #159985152] Setup dummy data for tests in the base.py and Re-write tests for new conditions
